feat: extract sorted-type-sets package; grades are now canonical (sorted) sets via Generic-derived ordering
