--- src/Control/Monad/Graded/Grade.hs
@@ -1,150 +1,0 @@
-{-# LANGUAGE AllowAmbiguousTypes #-}
-{-# LANGUAGE ConstraintKinds #-}
-{-# LANGUAGE DataKinds #-}
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE KindSignatures #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE PolyKinds #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE TypeFamilies #-}
-{-# LANGUAGE TypeOperators #-}
-{-# LANGUAGE UndecidableInstances #-}
-
--- | The grade set-algebra for errors: membership and the subset injection that
--- embeds a smaller error sum into a larger one.  Phrased over the injective
--- 'Tensored' newtype (not the non-injective 'MConcat') so the list indices stay
--- inferable, and driven by a closed-family position witness so nothing relies on
--- overlapping instances.
-module Control.Monad.Graded.Grade
-  ( Member,
-    Handle,
-    inj,
-    decompose,
-    Subset (..),
-    Union,
-    Nub,
-    Remove,
-    Delete,
-    here,
-    there,
-    uncons,
-  )
-where
-
---------------------------------------------------------------------------------
-
-import Control.Category.Tensor.Expr (Tensored (..), type (++))
-import Data.Kind (Type)
-import Data.Void (Void, absurd)
-
---------------------------------------------------------------------------------
--- Set algebra on grades (closed type families).
-
-type family Remove (x :: Type) (xs :: [Type]) :: [Type] where
-  Remove _ '[] = '[]
-  Remove x (x ': xs) = Remove x xs
-  Remove x (y ': xs) = y ': Remove x xs
-
-type family Nub (xs :: [Type]) :: [Type] where
-  Nub '[] = '[]
-  Nub (x ': xs) = x ': Nub (Remove x xs)
-
--- | Set union of two grades: concatenate then dedup (first-occurrence order).
-type family Union (xs :: [Type]) (ys :: [Type]) :: [Type] where
-  Union xs ys = Nub (xs ++ ys)
-
--- | Remove the element at a given position (purely structural — no apartness).
-type family DeleteAt (n :: Nat) (xs :: [Type]) :: [Type] where
-  DeleteAt 'Z (x ': xs) = xs
-  DeleteAt ('S n) (y ': xs) = y ': DeleteAt n xs
-
--- | Remove the first occurrence of a type from a grade.  On a deduped (set)
--- grade this removes its single occurrence.
-type Delete e es = DeleteAt (FindElem e es) es
-
---------------------------------------------------------------------------------
-
--- Injective helpers: signatures pin the list indices through 'Tensored'.
-
-here :: e -> Tensored Either Void (e ': es)
-here = Tensored . Left
-
-there :: Tensored Either Void es -> Tensored Either Void (f ': es)
-there = Tensored . Right . getTensored
-
-uncons :: Tensored Either Void (x ': xs) -> Either x (Tensored Either Void xs)
-uncons t = case getTensored t of
-  Left x -> Left x
-  Right r -> Right (Tensored r)
-
---------------------------------------------------------------------------------
-
--- Closed-family position witness (no overlapping instances).
-
-data Nat = Z | S Nat
-
-type family FindElem (x :: Type) (xs :: [Type]) :: Nat where
-  FindElem x (x ': xs) = 'Z
-  FindElem x (y ': xs) = 'S (FindElem x xs)
-
-class AtIndex (n :: Nat) (x :: Type) (xs :: [Type]) where
-  injAt :: x -> Tensored Either Void xs
-
-instance AtIndex 'Z x (x ': xs) where
-  injAt = here
-
-instance (AtIndex n x xs) => AtIndex ('S n) x (y ': xs) where
-  injAt = there . injAt @n @x @xs
-
--- | @x@ is a member of the type-set @xs@.
-type Member x xs = AtIndex (FindElem x xs) x xs
-
--- | Inject a single member into the n-ary error sum at its position.
-inj :: forall x xs. (Member x xs) => x -> Tensored Either Void xs
-inj = injAt @(FindElem x xs) @x @xs
-
---------------------------------------------------------------------------------
-
--- | @xs@ is a subset of @ys@; embed the smaller error sum into the larger.
-class Subset xs ys where
-  injSub :: Tensored Either Void xs -> Tensored Either Void ys
-
-instance Subset '[] ys where
-  injSub = absurd . getTensored
-
-instance (Member x ys, Subset xs ys) => Subset (x ': xs) ys where
-  injSub t = case uncons t of
-    Left x -> inj x
-    Right r -> injSub r
-
---------------------------------------------------------------------------------
--- Decompose: split an error sum at a given member (used by narrowing catch).
-
-class DecomposeAt (n :: Nat) (e :: Type) (es :: [Type]) where
-  decomposeAt :: Tensored Either Void es -> Either e (Tensored Either Void (DeleteAt n es))
-
-instance DecomposeAt 'Z e (e ': es) where
-  decomposeAt t = case uncons t of
-    Left e -> Left e
-    Right r -> Right r
-
-instance (DecomposeAt n e es) => DecomposeAt ('S n) e (y ': es) where
-  decomposeAt t = case uncons t of
-    Left y -> Right (here y)
-    Right r -> case decomposeAt @n @e r of
-      Left e -> Left e
-      Right r' -> Right (there r')
-
--- | The constraint enabling narrowing catch on error @e@ within grade @es@.
-type Handle e es = DecomposeAt (FindElem e es) e es
-
--- | Project a member @e@ out of an error sum, or return the remaining sum with
--- @e@ removed.
-decompose ::
-  forall e es.
-  (Handle e es) =>
-  Tensored Either Void es ->
-  Either e (Tensored Either Void (Delete e es))
-decompose = decomposeAt @(FindElem e es) @e @es
